https://github.com/blockscout/blockscout/issues/4522

## Motivation

Query for `uncataloged_token_transfer_block_numbers` method is slow (> 60 mins on some application instances).

## Changelog

- Remove join with `transactions` table since `block_hash` column is directly available at `logs` table.
- Replace `LEFT JOIN  / IS NULL` approach for finding records in table A, that do not exist in table B, with `NOT EXISTS` approach, which should be faster in Postgres.

`EXPLAIN ANALYZE` for the query before the changes:
```
EXPLAIN ANALYZE SELECT DISTINCT
ON (t1."block_number") t1."block_number"
FROM "logs" AS l0 INNER JOIN "transactions" AS t1 ON t1."hash" = l0."transaction_hash"
LEFT OUTER JOIN "token_transfers" AS t2 ON (t2."transaction_hash" = l0."transaction_hash") AND (t2."log_index" = l0."index")
WHERE (l0."first_topic" = '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef')
AND ((t2."transaction_hash" IS NULL) AND (t2."log_index" IS NULL))
AND (NOT (t1."block_hash" IS NULL));
                                                                          QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=129460.16..129804.55 rows=23858 width=4) (actual time=512.629..517.434 rows=0 loops=1)
   ->  Sort  (cost=129460.16..129632.35 rows=68879 width=4) (actual time=512.628..517.432 rows=0 loops=1)
         Sort Key: t1.block_number
         Sort Method: quicksort  Memory: 25kB
         ->  Gather  (cost=113081.24..123925.12 rows=68879 width=4) (actual time=512.604..517.407 rows=0 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               ->  Parallel Hash Join  (cost=112081.24..116037.22 rows=28700 width=4) (actual time=497.722..497.727 rows=0 loops=3)
                     Hash Cond: (l0.transaction_hash = t1.hash)
                     ->  Merge Anti Join  (cost=86494.15..88958.66 rows=31416 width=33) (actual time=397.292..397.294 rows=0 loops=3)
                           Merge Cond: ((l0.transaction_hash = t2.transaction_hash) AND (l0.index = t2.log_index))
                           ->  Sort  (cost=51693.93..51928.16 rows=93691 width=37) (actual time=131.780..139.406 rows=73661 loops=3)
                                 Sort Key: l0.transaction_hash, l0.index
                                 Sort Method: external merge  Disk: 3688kB
                                 Worker 0:  Sort Method: external merge  Disk: 3544kB
                                 Worker 1:  Sort Method: external merge  Disk: 3608kB
                                 ->  Parallel Seq Scan on logs l0  (cost=0.00..41395.10 rows=93691 width=37) (actual time=0.046..79.3
86 rows=73661 loops=3)
                                       Filter: ((first_topic)::text = '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523
b3ef'::text)
                                       Rows Removed by Filter: 172442
                           ->  Sort  (cost=34800.22..35355.68 rows=222183 width=37) (actual time=190.408..221.526 rows=221197 loops=3
)
                                 Sort Key: t2.transaction_hash, t2.log_index
                                 Sort Method: external merge  Disk: 10840kB
                                 Worker 0:  Sort Method: external merge  Disk: 10840kB
                                 Worker 1:  Sort Method: external merge  Disk: 10840kB
                                 ->  Seq Scan on token_transfers t2  (cost=0.00..8992.83 rows=222183 width=37) (actual time=0.052..48
.719 rows=221199 loops=3)
                     ->  Parallel Hash  (cost=23203.94..23203.94 rows=117292 width=37) (actual time=83.041..83.042 rows=93988 loops=3
)
                           Buckets: 65536  Batches: 8  Memory Usage: 3072kB
                           ->  Parallel Seq Scan on transactions t1  (cost=0.00..23203.94 rows=117292 width=37) (actual time=0.156..5
1.997 rows=93988 loops=3)
                                 Filter: (block_hash IS NOT NULL)
                                 Rows Removed by Filter: 8727
 Planning Time: 2.094 ms
 Execution Time: 521.339 ms
```

`EXPLAIN ANALYZE for the query after the changes`:

```
EXPLAIN ANALYZE SELECT DISTINCT
ON (l0."block_number") l0."block_number"
FROM "logs" AS l0
WHERE (l0."first_topic" = '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef')
AND l0."block_hash" IS NOT NULL
AND NOT EXISTS (
SELECT 1 FROM "token_transfers" AS t2
WHERE (t2."transaction_hash" = l0."transaction_hash") AND (t2."log_index" = l0."index")
);
                                                                                QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Unique  (cost=103928.52..104305.51 rows=20339 width=4) (actual time=472.654..475.359 rows=0 loops=1)
   ->  Sort  (cost=103928.52..104117.02 rows=75398 width=4) (actual time=472.653..475.358 rows=0 loops=1)
         Sort Key: l0.block_number
         Sort Method: quicksort  Memory: 25kB
         ->  Gather  (cost=87816.15..97820.44 rows=75398 width=4) (actual time=472.648..475.351 rows=0 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               ->  Merge Anti Join  (cost=86816.15..89280.64 rows=31416 width=4) (actual time=457.809..457.810 rows=0 loops=3)
                     Merge Cond: ((l0.index = t2.log_index) AND (l0.transaction_hash = t2.transaction_hash))
                     ->  Sort  (cost=52015.93..52250.16 rows=93691 width=41) (actual time=153.101..164.087 rows=73661 loops=3)
                           Sort Key: l0.index, l0.transaction_hash
                           Sort Method: external merge  Disk: 4152kB
                           Worker 0:  Sort Method: external merge  Disk: 3776kB
                           Worker 1:  Sort Method: external merge  Disk: 3776kB
                           ->  Parallel Seq Scan on logs l0  (cost=0.00..41395.10 rows=93691 width=41) (actual time=0.135..92.236 row
s=73661 loops=3)
                                 Filter: ((block_hash IS NOT NULL) AND ((first_topic)::text = '0xddf252ad1be2c89b69c2b068fc378daa952b
a7f163c4a11628f55a4df523b3ef'::text))
                                 Rows Removed by Filter: 172442
                     ->  Sort  (cost=34800.22..35355.68 rows=222183 width=37) (actual time=210.823..255.975 rows=221194 loops=3)
                           Sort Key: t2.log_index, t2.transaction_hash
                           Sort Method: external merge  Disk: 10840kB
                           Worker 0:  Sort Method: external merge  Disk: 10840kB
                           Worker 1:  Sort Method: external merge  Disk: 10840kB
                           ->  Seq Scan on token_transfers t2  (cost=0.00..8992.83 rows=222183 width=37) (actual time=0.050..47.349 r
ows=221199 loops=3)
 Planning Time: 1.096 ms
 Execution Time: 479.153 ms
```



## Checklist for your Pull Request (PR)

<!--
  Ideally a PR has all of the checkmarks set.

  If something in this list is irrelevant to your PR, you should still set this
  checkmark indicating that you are sure it is dealt with (be that by irrelevance).

  If you don't set a checkmark (e. g. don't add a test for new functionality),
  please justify why.
-->

  - [x] I added an entry to `CHANGELOG.md` with this PR
  - [ ] If I added new functionality, I added tests covering it.
  - [ ] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
  - [x] I checked whether I should update the docs and did so by submitting a PR to https://github.com/blockscout/docs
  - [x] If I added/changed/removed ENV var, I submitted a PR to https://github.com/blockscout/docs to update the list of env vars at https://github.com/blockscout/docs/blob/master/for-developers/information-and-settings/env-variables.md and I updated the version to `master` in the Version column. Changes will be reflected in this table: https://docs.blockscout.com/for-developers/information-and-settings/env-variables. 
  - [x] If I add new indices into DB, I checked, that they are not redundant with PGHero or other tools